WPA-SEC download into .pcap.cracked single files
This commit implements the "single_files" setting for the WPA-SEC plugin, which already existed for the Onlinehashcrack plugin.

This option, if set to true, downloads the cracked passwords from the wpasec site into individual files with the .pcap.cracked extension, so you can see the cracked WiFi passwords directly in the webgpsmap plugin map.

Additionally, this commit fixes a crash that occurred if the "download_interval" option was not set, falling back to a value of 3600.

Furthermore, while rewriting the code I improved the log messages and exception handling (for example, by calling raise_for_status() after making HTTP requests and using the logging.exception() method, which prints the exception stacktrace to the logs for easier debugging).